+++ prover/prover_fri/src/main.rs
@@ -35,7 +35,9 @@ async fn graceful_shutdown(port: u16) -> anyhow::Result<impl Future<Output = ()>
     let zone = get_zone().await.context("get_zone()")?;
     let address = SocketAddress { host, port };
     Ok(async move {
-        pool.access_storage().await.unwrap()
+        pool.access_storage()
+            .await
+            .unwrap()
             .fri_gpu_prover_queue_dal()
             .update_prover_instance_status(address, GpuProverInstanceStatus::Dead, zone)
             .await
@@ -67,8 +69,7 @@ async fn main() -> anyhow::Result<()> {
         tracing::info!("No sentry URL was provided");
     }
 
-    let prover_config = FriProverConfig::from_env()
-        .context("FriProverConfig::from_env()")?;
+    let prover_config = FriProverConfig::from_env().context("FriProverConfig::from_env()")?;
     let exporter_config = PrometheusExporterConfig::pull(prover_config.prometheus_port);
 
     let (stop_signal_sender, stop_signal_receiver) = oneshot::channel();
@@ -81,24 +82,24 @@ async fn main() -> anyhow::Result<()> {
     .context("Error setting Ctrl+C handler")?;
 
     let (stop_sender, stop_receiver) = tokio::sync::watch::channel(false);
-    let blob_store = ObjectStoreFactory::prover_from_env()
-        .context("ObjectStoreFactory::prover_from_env()")?;
+    let blob_store =
+        ObjectStoreFactory::prover_from_env().context("ObjectStoreFactory::prover_from_env()")?;
     let public_blob_store = match prover_config.shall_save_to_public_bucket {
         false => None,
         true => Some(
             ObjectStoreFactory::new(
                 ObjectStoreConfig::public_from_env()
-                    .context("ObjectStoreConfig::public_from_env()")?
+                    .context("ObjectStoreConfig::public_from_env()")?,
             )
-                .create_store()
-                .await,
+            .create_store()
+            .await,
         ),
     };
     let specialized_group_id = prover_config.specialized_group_id;
 
     let circuit_ids_for_round_to_be_proven = FriProverGroupConfig::from_env()
         .get_circuit_ids_for_group_id(specialized_group_id)
-        .unwrap_or(vec![]);
+        .unwrap_or_default();
     let circuit_ids_for_round_to_be_proven =
         get_all_circuit_id_round_tuples_for(circuit_ids_for_round_to_be_proven);
 
@@ -166,8 +167,8 @@ async fn get_prover_tasks(
         vk_commitments
     );
 
-    let setup_load_mode = load_setup_data_cache(&prover_config)
-        .context("load_setup_data_cache()")?;
+    let setup_load_mode =
+        load_setup_data_cache(&prover_config).context("load_setup_data_cache()")?;
     let prover = Prover::new(
         store_factory.create_store().await,
         public_blob_store,
@@ -195,8 +196,8 @@ async fn get_prover_tasks(
     use tokio::sync::Mutex;
     use zksync_prover_fri_types::queue::FixedSizeQueue;
 
-    let setup_load_mode = gpu_prover::load_setup_data_cache(&prover_config)
-        .context("load_setup_data_cache()")?;
+    let setup_load_mode =
+        gpu_prover::load_setup_data_cache(&prover_config).context("load_setup_data_cache()")?;
     let witness_vector_queue = FixedSizeQueue::new(prover_config.queue_capacity);
     let shared_witness_vector_queue = Arc::new(Mutex::new(witness_vector_queue));
     let consumer = shared_witness_vector_queue.clone();